refactor(vulkan): extract device discovery into separate module
Rollback trigger: test failures in ramshared-vulkan
